package maintainer refactor
