+++ src/main/java/com/ethosengine/sonarqube/auth/oidc/AutoLoginFilter.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.vaulttec.sonarqube.auth.oidc;
+package com.ethosengine.sonarqube.auth.oidc;
 
 import java.io.IOException;
