+++ src/main/java/com/ethosengine/sonarqube/auth/oidc/package-info.java
@@ -15,20 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.vaulttec.sonarqube.auth.oidc;
+@ParametersAreNonnullByDefault
+package com.ethosengine.sonarqube.auth.oidc;
 
-import org.sonar.api.Plugin;
-import org.sonar.api.SonarQubeSide;
-
-public class AuthOidcPlugin implements Plugin {
-
-  @Override
-  public void define(Context context) {
-    if (context.getRuntime().getSonarQubeSide() == SonarQubeSide.SERVER) {
-      context.addExtensions(OidcConfiguration.class, OidcClient.class, OidcIdentityProvider.class,
-          UserIdentityFactory.class, AutoLoginFilter.class);
-      context.addExtensions(OidcConfiguration.definitions());
-    }
-  }
-
-}
+import javax.annotation.ParametersAreNonnullByDefault;
